+++ bfabric_app_runner/src/bfabric_app_runner/actions/execute.py
@@ -1,7 +1,10 @@
 from pathlib import Path
-from typing import assert_never
+from typing import assert_never, Any
+
+from loguru import logger
 
 from bfabric import Bfabric
+from bfabric.entities import WorkflowTemplateStep, WorkflowTemplate, WorkflowStep
 from bfabric.experimental.entity_lookup_cache import EntityLookupCache
 from bfabric.experimental.workunit_definition import WorkunitDefinition
 from bfabric_app_runner.actions.types import (
@@ -38,8 +41,7 @@ def execute(action: ActionGeneric, client: Bfabric) -> None:
 
 def execute_dispatch(action: ActionDispatch, client: Bfabric) -> None:
     """Executes a dispatch action."""
-
-    app_version, _, workunit_ref = load_workunit_information(
+    app_version, bfabric_app_spec, workunit_ref = load_workunit_information(
         app_spec=action.app_ref, client=client, work_dir=action.work_dir, workunit_ref=action.workunit_ref
     )
 
@@ -49,8 +51,17 @@ def execute_dispatch(action: ActionDispatch, client: Bfabric) -> None:
 
     if not action.read_only:
         # Set the workunit status to processing
-        workunit_definition = WorkunitDefinition.from_yaml(action.work_dir / "workunit_definition.yml")
+        workunit_definition = WorkunitDefinition.from_yaml(workunit_ref)
         client.save("workunit", {"id": workunit_definition.registration.workunit_id, "status": "processing"})
+
+        # Create a workflowstep template if specified
+        if bfabric_app_spec.workflow_template_step_id:
+            logger.info(f"Creating workflowstep from template {bfabric_app_spec.workflow_template_step_id}")
+            _register_workflow_step(
+                workflow_template_step_id=bfabric_app_spec.workflow_template_step_id,
+                workunit_definition=workunit_definition,
+                client=client,
+            )
 
 
 def execute_run(action: ActionRun, client: Bfabric) -> None:
@@ -117,6 +128,64 @@ def execute_outputs(action: ActionOutputs, client: Bfabric) -> None:
                 force_storage=action.force_storage,
                 reuse_default_resource=True,
             )
+
+
+def _register_workflow_step(
+    workflow_template_step_id: int, workunit_definition: WorkunitDefinition, client: Bfabric
+) -> None:
+    # Load the workflow template step
+    workflow_template_step = WorkflowTemplateStep.find(id=workflow_template_step_id, client=client)
+    if not workflow_template_step:
+        logger.error(f"Misconfigured {workflow_template_step_id=}, cannot find it in the database.")
+        return
+
+    # Find or create the workflow entity
+    workflow = _find_or_create_workflow(
+        workflow_template=workflow_template_step.workflow_template,
+        workunit_definition=workunit_definition,
+        client=client,
+    )
+
+    # Prepare the data for the workflow step
+    workunit_id = workunit_definition.registration.workunit_id
+    workflow_step_data = {
+        "workflowid": workflow["id"],
+        "workflowtemplatestepid": workflow_template_step.id,
+        "supervisorid": workunit_definition.registration.user_id,
+        "workunitid": workunit_id,
+    }
+    if workunit_definition.execution.dataset is not None:
+        workflow_step_data["datasetid"] = workunit_definition.execution.dataset
+
+    # Find or create the workflow step entity
+    _create_workflow_step_if_not_exists(
+        workflow=workflow, workflow_step_data=workflow_step_data, workunit_id=workunit_id, client=client
+    )
+
+
+def _create_workflow_step_if_not_exists(
+    workflow: dict[str, Any], workflow_step_data: dict[str, Any], workunit_id: int, client: Bfabric
+) -> None:
+    workflow_step = WorkflowStep.find_by(workflow_step_data, client=client)
+    if workflow_step:
+        logger.info(f"Workflow step already exists: {workflow_step}, skipping creation.")
+        return
+    resp = client.save("workflowstep", workflow_step_data)
+    logger.info(f"Created workflow step: {resp[0]['id']} for workunit {workunit_id} in workflow {workflow['id']}.")
+
+
+def _find_or_create_workflow(
+    workflow_template: WorkflowTemplate, workunit_definition: WorkunitDefinition, client: Bfabric
+) -> dict[str, Any]:
+    container_id = workunit_definition.registration.container_id
+    workflow_data = {"containerid": container_id, "workflowtemplateid": workflow_template.id}
+    resp = client.read("workflow", workflow_data)
+    if len(resp) > 0:
+        workflow = resp[0]
+    else:
+        resp = client.save("workflow", workflow_data)
+        workflow = resp[0]
+    return workflow
 
 
 def _validate_chunks_list(work_dir: Path, chunk: str | None) -> list[Path]: