+++ bfabric_app_runner/src/bfabric_app_runner/dispatch/dispatch_single_resource_flow.py
@@ -1,13 +1,13 @@
+from pathlib import Path
+
+from loguru import logger
+from pydantic import BaseModel, ConfigDict
+
 from bfabric import Bfabric
 from bfabric.entities import Resource
 from bfabric.experimental.workunit_definition import WorkunitDefinition
-from loguru import logger
-from pathlib import Path
-from pydantic import BaseModel, ConfigDict
-
 from bfabric_app_runner.dispatch.generic import (
     write_chunks_file,
-    write_workunit_definition_file,
 )
 from bfabric_app_runner.dispatch.resource_flow import get_resource_flow_input_resources
 
@@ -35,6 +35,5 @@ class DispatchSingleResourceFlow:
             definition=definition,
             filter_suffix=self._config.filter_suffix,
         )
-        write_workunit_definition_file(out_dir=self._out_dir, definition=definition)
         path = self.dispatch_job(resources=resources, workunit=definition)
         write_chunks_file(out_dir=self._out_dir, chunks=[path])